+++ lib/all_widgets.dart
@@ -1,8 +1,2 @@
 library all_widgets;
 export 'package:all_widgets/export.dart';
-
-/// A Calculator.
-class Calculator {
-  /// Returns [value] plus 1.
-  int addOne(int value) => value + 1;
-}