+++ lib/custom_text/c_text.dart
@@ -3,11 +3,6 @@ import 'package:all_widgets/export.dart';
 class CustomText extends StatelessWidget {
   final String? text;
   final TextStyle? style;
-  final double? fontSize;
-  final Color? fontColor;
-  final Color? backgroundColor;
-  final TextDecoration? fontDecoration;
-  final FontWeight? fontWeight;
   final int? maxLines;
   final TextAlign? textAlign;
   final TextOverflow? overflow;
@@ -16,12 +11,7 @@ class CustomText extends StatelessWidget {
     super.key,
     this.text,
     this.style,
-    this.fontSize,
-    this.fontWeight,
     this.maxLines,
-    this.fontColor,
-    this.fontDecoration,
-    this.backgroundColor,
     this.textAlign,
     this.overflow,
   });