+++ CHANGELOG.md
@@ -1,3 +1,17 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* This Release contain now only one custom widget that is CustomText.
+
+
+Here is Widget tree.
+
+all_widgets/
+├── lib/
+│   ├── all_widgets.dart
+│   └── custom_text_field.dart
+├── test/
+│   └── custom_text_field_test.dart
+├── pubspec.yaml
+├── README.md
+├── CHANGELOG.md
+└── LICENSE